+++ src/basis/utils.jl
@@ -1,147 +1,3 @@
-
-## Utilities
-
-function unique(b::AbstractArray{Num}, simplify_eqs::Bool)
-    b = simplify_eqs ? simplify.(b) : b
-    returns = ones(Bool, size(b)...)
-    N = maximum(eachindex(b))
-    for i ∈ eachindex(b)
-        returns[i] = !any([isequal(b[i], b[j]) for j in i+1:N])
-    end
-    return Num.(b[returns])
-end
-
-
-function Base.unique!(b::AbstractArray{Num}, simplify_eqs = false)
-    bs = simplify_eqs ? simplify.(b) : b
-    removes = zeros(Bool, size(bs)...)
-    N = maximum(eachindex(bs))
-    for i ∈ eachindex(bs)
-        removes[i] = any([isequal(bs[i], bs[j]) for j in i+1:N])
-    end
-    deleteat!(b, removes)
-end
-
-function unique(b::AbstractArray{Equation}, simplify_eqs::Bool)
-    b = simplify_eqs ? simplify.(b) : b
-    returns = ones(Bool, size(b)...)
-    N = maximum(eachindex(b))
-    for i ∈ eachindex(b)
-        returns[i] = !any([isequal(b[i].rhs, b[j].rhs) for j in i+1:N])
-    end
-    return b[returns]
-end
-
-function Base.unique!(b::AbstractArray{Equation}, simplify_eqs::Bool)
-    bs = [bi.rhs for bi in b]
-    bs = simplify_eqs ? simplify.(bs) : bs
-    removes = zeros(Bool, size(bs)...)
-    N = maximum(eachindex(bs))
-    for i ∈ eachindex(bs)
-        removes[i] = any([isequal(bs[i], bs[j]) for j in i+1:N])
-    end
-    deleteat!(b, removes)
-end
-
-## Interfacing && merging
-
-function Base.unique!(b::Basis, simplify_eqs = false; eval_expression = false)
-    unique!(b.eqs, simplify_eqs)
-    update!(b, eval_expression)
-end
-
-function Base.unique(b::Basis; kwargs...)
-    eqs = unique(equations(b))
-    return Basis(eqs, states(b), parameters = parameters(b), iv = independent_variable(b), kwargs...)
-end
-
-"""
-    deleteat!(basis, inds, eval_expression = false)
-
-    Delete the entries specified by `inds` and update the `Basis` accordingly.
-"""
-function Base.deleteat!(b::Basis, inds; eval_expression = false)
-    deleteat!(b.eqs, inds)
-    update!(b, eval_expression)
-    return
-end
-
-"""
-    push!(basis, eq, simplify_eqs = true; eval_expression = false)
-
-    Push the equations(s) in `eq` into the basis and update all internal fields accordingly.
-    `eq` can either be a single equation or an array. If `simplify_eq` is true, the equation will be simplified.
-"""
-function Base.push!(b::Basis, eqs::AbstractArray, simplify_eqs = true; eval_expression = false)
-    @inbounds for eq ∈ eqs
-        push!(b, eq, false)
-    end
-    unique!(b, simplify_eqs, eval_expression = eval_expression)
-    return
-end
-
-function Base.push!(b::Basis, eq, simplify_eqs = true; eval_expression = false)
-    push!(equations(b), Variable(:φ, length(b.eqs)+1)~eq)
-    unique!(b, simplify_eqs, eval_expression = eval_expression)
-    return
-end
-
-function Base.push!(b::Basis, eq::Equation, simplify_eqs = true; eval_expression = false)
-    push!(equations(b), eq)
-    unique!(b, simplify_eqs, eval_expression = eval_expression)
-    return
-end
-
-
-"""
-    merge(x::Basis, y::Basis; eval_expression = false)
-
-    Return a new `Basis`, which is defined via the union of `x` and `y` .
-"""
-function Base.merge(x::Basis, y::Basis; eval_expression = false)
-    b =  unique(vcat([xi.rhs  for xi ∈ equations(x)], [xi.rhs  for xi ∈ equations(y)]))
-    vs = unique(vcat(states(x), states(y)))
-    ps = unique(vcat(parameters(x), parameters(y)))
-
-    c = unique(vcat(controls(x), controls(y)))
-    observed = unique(vcat(get_observed(x), get_observed(y)))
-
-    return Basis(Num.(b), vs, parameters = ps, controls = c, observed = observed, eval_expression = eval_expression)
-end
-
-"""
-    merge!(x::Basis, y::Basis; eval_expression = false)
-
-    Updates `x` to include the union of both `x` and `y`.
-"""
-function Base.merge!(x::Basis, y::Basis; eval_expression = false)
-    push!(x, map(x->x.rhs, equations(y)))
-    Core.setfield!(x, :states, unique(vcat(states(x),states(y))))
-    Core.setfield!(x, :ps, unique(vcat(parameters(x), parameters(y))))
-    Core.setfield!(x, :controls, unique(vcat(controls(x), controls(y))))
-    Core.setfield!(x, :observed, unique(vcat(get_observed(x), get_observed(y))))
-    update!(x, eval_expression)
-    return
-end
-
-## Additional functionalities
-
-function (==)(x::Basis, y::Basis)
-    length(x) == length(y) || return false
-    n = zeros(Bool, length(x))
-    yrhs = [yi.rhs for yi in equations(y)]
-    xrhs = [xi.rhs for xi in equations(x)]
-    @inbounds for (i, xi) in enumerate(xrhs)
-        n[i] = any(isequal.([xi], yrhs))
-        !n[i] && break
-    end
-    return all(n)
-end
-
-free_parameters(b::AbstractBasis; operations = [+]) = count_operation([xi.rhs for xi in b.eqs], operations) + length(b.eqs)
-
-##
-
 ## Get unary and binary functions
 
 function is_unary(f::Function, t::Type = Number)
